调整速度和扭矩限制修改逻辑，修改cyc为uint64

+++ drivers/motor/dji/motor_dji.h
@@ -72,8 +72,8 @@ struct dji_motor_data {
 	int8_t RAWtemp;
 	int32_t angle_add;
 
-	uint32_t curr_time;
-	uint32_t prev_time;
+	uint64_t curr_time;
+	uint64_t prev_time;
 	int8_t missed_times;
 
 	float angle_offset;
